Copy the ItemStack regardless of its contents
When iterating over a table of inventory slots in lua, the iteration
operation ends when null is encountered, as there is no conversion from
null to nil. Directly referencing a table element that represents an
empty inventory slot also causes a traceback to occur.

This commit copies all ItemStacks, even empty ones, which are in fact
ItemStack.EMPTY, and can be identified by the name minecraft:air in lua
programs. Doing this allows iteration over a table representing an
inventory to complete successfully, and direct referencing of all table
elements is possible.